Enhanced file and directory components

+++ src/Qweree.WebApplication/Infrastructure/View/Text/Plurality.cs
@@ -2,7 +2,7 @@ namespace Qweree.WebApplication.Infrastructure.View.Text
 {
     public static class Plurality
     {
-        public static string Decide(int value, string singular, string plural)
+        public static string Decide(long value, string singular, string plural)
         {
             if (value == 1)
                 return singular;
